+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.4.2]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [3.5.0]
+This version adds support for globus_sdk v4.4.0 and pandas v3.0, and gracefully handles Alyx request rate limits.
+
+### Modified
+
+- support latest version of Globus SDK
+- requirements moved to pyproject.toml
+- support pandas 3.0
+- AlyxClient will now wait before retrying REST requests that return a 429 or 503 status
+
+### Fixed
+
+- registration.get_dataset_type now casts dtypes input to 
+- as_dict=False no longer ignored for list inputs to ConversionMixin.path2ref
+
+## [3.4.2]
 This version fixes the saving and loading of insertions parquet tables for offline processing.
 
 ### Fixed